Get CommonJS style to work in Node.js
Grunt Dust supports to module systems, AMD and CommonJS. CommonJS
is the basis of RequireJS. One of the odd things though, is that
it  doens't return an object. So, when using it in Node.js, you
don't get back the Dust template cache.

Added in the ability to pass in `packageName` as an option that
will define what variable to return. More than likely it will be
the value for the dust variable. But, instead of trying to guess
it, I've made it explicit.